refactor(mentor): 콤포넌트명과 파일명 일치 하도록 변경
